Create caesar_cipher
The input to the program is a line of text in English, in which we need to encrypt all the words. Each word in the string should be encrypted using the Caesar cipher (cyclic shift by the length of that word). Lowercase letters remain lowercase, while uppercase letters remain uppercase.